Add files via upload
The provided Java program calculates the number of trailing zeros in the factorial of a given number n m . Trailing zeros in a factorial result from factors of 10 in the number. Each 10 is composed of a factor of 2 and a factor of 5, and since there are always more factors of 2 than factors of 5, the number of trailing zeros is determined by the number of factors of 5.